% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNESS.r
\name{CNESS}
\alias{CNESS}
\title{The Chord-distance Normalized Expected Specis Shared (CNESS)}
\usage{
CNESS(X, m)
}
\arguments{
\item{X}{Community data matrix, samples as rows, species as column}

\item{m}{NESSm value, can range from 1 (high weight to adundant species) to the minimum sample total (high weight to rare species). The NESSm function compute the best trade-off.}
}
\value{
Standardized data frame
}
\description{
The function converts raw community data to hypergeometric probabilities from a random draw of NESSm individuals followed by station normalization. The CNESS distance can then be obtained by computing the Euclidean distance on the transformed data (see Trueblood et al., 1994; Gallagher 1996).
}
\examples{
library(vegan)
data(varespec)
m <- NESSm(varespec)
cness.mat <- CNESS(varespec, m)
cness.dist <- dist(cness.mat, "euclidean")

}
\references{
Gallagher, E.D., 1996. COMPAH documentation. 65 p. http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.9.1334&rep=rep1&type=pdf.
Trueblood, D.D., Gallagher, E.D., Gould, D.M., 1994. Three stages of seasonal succession on the Savin Hill Cove mudflat, Boston Harbor. Limnology and Oceanography 39, 1440-1454.
}
