% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ness.R
\docType{package}
\name{ness}
\alias{ness}
\alias{ness-package}
\title{ness Package}
\description{
Hypergeometric standardization and probabilistic measures of similarity for community ecology
}
\author{
Lénaick Menot \email{lenaick.menot@ifremer.fr}
}
