% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NESSm.R
\name{NESSm}
\alias{NESSm}
\title{NESSm sample size}
\usage{
NESSm(X)
}
\arguments{
\item{X}{Community data matrix, samples as rows, species as column}
}
\value{
sample size to be used in CNESS or NNESS to produce an index sensitive to both rare and abundant species
}
\description{
Computes the sample size that provides the best trade-off between NESSm = 1 (giving high weight to abundant species) and NESSm= "minimum sample total" (giving high weight to rare species), after Trueblood et al. (1994) and Gallagher (1996).
}
\examples{
require(vegan)
data(varespec)
m <- NESSm(varespec) # Be patient, computation is slow

### Note that the function doesn't work with large numbers nor data matrix with no variance
largenumber <- matrix(rnorm(100, 1000, 100), 10 , 10)
m <- NESSm(largenumber)

}
\references{
Gallagher, E.D., 1996. COMPAH documentation. 65 p. http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.9.1334&rep=rep1&type=pdf.
Trueblood, D.D., Gallagher, E.D., Gould, D.M., 1994. Three stages of seasonal succession on the Savin Hill Cove mudflat, Boston Harbor. Limnology and Oceanography 39, 1440-1454.
}
