% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNESS.R
\name{NNESS}
\alias{NNESS}
\title{New Normalized Expected Specis Shared (NNESS)}
\usage{
NNESS(X, m)
}
\arguments{
\item{X}{Community data matrix, samples as rows, species as column}

\item{m}{NESSm value, can range from 1 (high weight to adundant species) to the minimum sample total (high weight to rare species). The NESSm function compute the best trade-off.}
}
\value{
distance object, the similarity matrix between samples

 @examples
library(vegan)
data(varespec)
m <- NESSm(varespec)
nness.dist <- NNESS(varespec, m)
}
\description{
This functions computes the NNESS similarity index from a random draw of NESSm individuals (see Gallagher, 1996).
}
\references{
Gallagher, E.D., 1996. COMPAH documentation. 65 p. http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.9.1334&rep=rep1&type=pdf.
}
